backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "procedural-game-asset-foundry"
version = "0.1.0"
description = "Production-grade JSON-native asset generation backend for game developers"
authors = [
    {name = "Game Asset Foundry Team"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    "pillow>=10.1.0",
    "numpy>=1.24.3",
    "aiofiles>=23.2.1",
    "httpx>=0.25.2",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true